refactor: JS 파일명 수정 및 svg 입력 방식 변경
