+++ index.html
@@ -11,7 +11,12 @@
   <div class="basic-display">
     <div class="top-container">
       <div class="top-container__title">
-        <img class="top-container__icon" src="src/img/newsStandLogo.svg" />
+        <svg width="24" class="top-container__icon" height="21" viewBox="0 0 24 21" fill="none"
+          xmlns="http://www.w3.org/2000/svg">
+          <path
+            d="M20 2.25V17C19.9998 17.1249 20.0463 17.2455 20.1305 17.3378C20.2147 17.4301 20.3304 17.4876 20.4548 17.4989C20.5793 17.5102 20.7034 17.4745 20.8028 17.3988C20.9023 17.3231 20.9697 17.2129 20.992 17.09L21 17V3.5H22.25C22.6892 3.50002 23.1123 3.66517 23.4354 3.96268C23.7585 4.26019 23.9579 4.6683 23.994 5.106L24 5.25V17.25C24.0001 18.0801 23.6824 18.8788 23.1123 19.4822C22.5422 20.0856 21.7628 20.448 20.934 20.495L20.75 20.5H3.25C2.41986 20.5001 1.62117 20.1824 1.01777 19.6123C0.414367 19.0422 0.0519987 18.2628 0.00500011 17.434L0 17.25V2.25C1.64046e-05 1.81081 0.165171 1.38768 0.46268 1.06461C0.760189 0.741536 1.1683 0.542139 1.606 0.506L1.75 0.5H18.25C18.6892 0.500016 19.1123 0.665171 19.4354 0.96268C19.7585 1.26019 19.9579 1.6683 19.994 2.106L20 2.25ZM7.747 9.504H5.25C4.81081 9.50402 4.38768 9.66917 4.06461 9.96668C3.74154 10.2642 3.54214 10.6723 3.506 11.11L3.5 11.254V13.75C3.50002 14.1892 3.66517 14.6123 3.96268 14.9354C4.26019 15.2585 4.6683 15.4579 5.106 15.494L5.25 15.5H7.747C8.18635 15.5 8.60961 15.3347 8.93271 15.037C9.25581 14.7392 9.4551 14.3309 9.491 13.893L9.496 13.75V11.254C9.49623 10.8146 9.33119 10.3913 9.03366 10.068C8.73613 9.74469 8.32787 9.54515 7.89 9.509L7.747 9.504ZM15.75 14H12.25L12.148 14.007C11.9685 14.0317 11.804 14.1206 11.6849 14.2571C11.5658 14.3937 11.5002 14.5688 11.5002 14.75C11.5002 14.9312 11.5658 15.1063 11.6849 15.2429C11.804 15.3794 11.9685 15.4683 12.148 15.493L12.25 15.5H15.75L15.852 15.493C16.0315 15.4683 16.196 15.3794 16.3151 15.2429C16.4342 15.1063 16.4998 14.9312 16.4998 14.75C16.4998 14.5688 16.4342 14.3937 16.3151 14.2571C16.196 14.1206 16.0315 14.0317 15.852 14.007L15.75 14ZM5.25 11.004H7.747C7.80319 11.004 7.85772 11.023 7.90181 11.0578C7.9459 11.0927 7.97696 11.1413 7.99 11.196L7.997 11.254V13.75C7.99691 13.8063 7.9778 13.861 7.94277 13.9051C7.90774 13.9492 7.85884 13.9802 7.804 13.993L7.747 14H5.25C5.19368 13.9999 5.13904 13.9808 5.09493 13.9458C5.05083 13.9107 5.01984 13.8618 5.007 13.807L5 13.75V11.254C4.99987 11.1975 5.01887 11.1426 5.05392 11.0983C5.08897 11.054 5.13799 11.0229 5.193 11.01L5.25 11.004ZM15.75 9.504H12.25L12.148 9.51C11.9677 9.53389 11.8023 9.62246 11.6824 9.75923C11.5626 9.896 11.4965 10.0717 11.4965 10.2535C11.4965 10.4353 11.5626 10.611 11.6824 10.7478C11.8023 10.8845 11.9677 10.9731 12.148 10.997L12.25 11.004H15.75L15.852 10.997C16.0323 10.9731 16.1977 10.8845 16.3176 10.7478C16.4374 10.611 16.5035 10.4353 16.5035 10.2535C16.5035 10.0717 16.4374 9.896 16.3176 9.75923C16.1977 9.62246 16.0323 9.53389 15.852 9.51L15.75 9.504ZM15.75 4.997H4.25L4.148 5.004C3.96849 5.0287 3.80399 5.11755 3.68491 5.25414C3.56583 5.39072 3.50023 5.5658 3.50023 5.747C3.50023 5.9282 3.56583 6.10328 3.68491 6.23986C3.80399 6.37645 3.96849 6.4653 4.148 6.49L4.25 6.497H15.75L15.852 6.49C16.0315 6.4653 16.196 6.37645 16.3151 6.23986C16.4342 6.10328 16.4998 5.9282 16.4998 5.747C16.4998 5.5658 16.4342 5.39072 16.3151 5.25414C16.196 5.11755 16.0315 5.0287 15.852 5.004L15.75 4.997Z"
+            fill="#4362D0" />
+        </svg>
         <span class="top-container__title-text">뉴스스탠드</span>
       </div>
       <div class="top-container__date">
@@ -44,11 +49,17 @@
         </li>
       </ul>
       <div class="press-container__view"></div>
-      <img class="press-container__left-arrow" src="src/img/LeftButton.svg" />
-      <img class="press-container__right-arrow" src="src/img/RightButton.svg" />
+      <svg class="press-container__left-arrow" width="26" height="42" viewBox="0 0 26 42" fill="none"
+        xmlns="http://www.w3.org/2000/svg">
+        <path d="M25 1L1 21L25 41" stroke="#6E8091" />
+      </svg>
+      <svg class="press-container__right-arrow" width="26" height="42" viewBox="0 0 26 42" fill="none"
+        xmlns="http://www.w3.org/2000/svg">
+        <path d="M1 41L25 21L1 1" stroke="#6E8091" />
+      </svg>
     </div>
   </div>
-  <script type="module" src="src/js/renderIndex.js"></script>
+  <script type="module" src="src/js/NewsStand.js"></script>
 </body>
 
 </html>
