+++ src/js/Crawler.js
@@ -1,0 +1,9 @@
+import crawlNews from "./NewsCrawler.js";
+import crawlLogos from "./LogoCrawler.js"
+
+const crawl = () => {
+  crawlLogos();
+  crawlNews();
+}
+
+crawl();
