+++ src/js/NewsTitleRenderer.js
