+++ src/js/PressTableRenderer.js
